@@ -2,7 +2,7 @@
 
     <PropertyGroup>
       <OutputType>WinExe</OutputType>
-      <TargetFramework>net9.0-windows</TargetFramework>
+      <TargetFramework>net10.0-windows</TargetFramework>
       <RootNamespace>QslLabel</RootNamespace>
       <UseWindowsForms>true</UseWindowsForms>
       <NeutralLanguage>en-US</NeutralLanguage>
